fix(transacao): trocar tipo de valor big decimal por double

+++ src/main/java/com/mauroraya/transacao_api/models/Transacao.java
@@ -1,9 +1,8 @@
 package com.mauroraya.transacao_api.models;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record Transacao(
-    BigDecimal valor,
+    double valor,
     OffsetDateTime dataHora
 ) {}
