+++ src/main/java/com/mauroraya/transacao_api/requests/CriarTransacaoRequest.java
@@ -1,6 +1,5 @@
 package com.mauroraya.transacao_api.requests;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 import jakarta.validation.constraints.NotNull;
@@ -9,7 +8,7 @@ import jakarta.validation.constraints.PositiveOrZero;
 
 public record CriarTransacaoRequest(
     @PositiveOrZero(message = "O valor deve ser maior ou igual a 0")
-    BigDecimal valor,
+    double valor,
 
     @NotNull(message = "Data e hora são obrigatórias")
     @PastOrPresent(message = "Data e hora não podem estar no futuro")
